# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Multi-Hospital Post-Discharge Outreach Platform",
    "description": "Phase 1 \u2014 Foundation: multi-tenant hospital management, JWT authentication, protocol management, tenant isolation, and audit logging.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/voice/sessions/{call_id}": {
      "get": {
        "tags": [
          "Voice"
        ],
        "summary": "Get Session Context",
        "description": "Called by the TypeScript LiveKit Agent to get the context for a call.",
        "operationId": "get_session_context_api_v1_voice_sessions__call_id__get",
        "parameters": [
          {
            "name": "call_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Call Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/voice/sessions/{call_id}/complete": {
      "post": {
        "tags": [
          "Voice"
        ],
        "summary": "Complete Session",
        "description": "Called by the TypeScript agent when the conversation ends.",
        "operationId": "complete_session_api_v1_voice_sessions__call_id__complete_post",
        "parameters": [
          {
            "name": "call_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Call Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConversationResult"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConversationOutcome": {
        "type": "string",
        "enum": [
          "COMPLETED",
          "CALLBACK_REQUESTED",
          "PATIENT_DECLINED",
          "INTERRUPTED",
          "TRANSFERRED",
          "NO_ANSWER"
        ],
        "title": "ConversationOutcome"
      },
      "ConversationResult": {
        "properties": {
          "call_id": {
            "type": "string",
            "format": "uuid",
            "title": "Call Id"
          },
          "patient_id": {
            "type": "string",
            "format": "uuid",
            "title": "Patient Id"
          },
          "campaign_id": {
            "type": "string",
            "format": "uuid",
            "title": "Campaign Id"
          },
          "outcome": {
            "$ref": "#/components/schemas/ConversationOutcome"
          },
          "patient_reached": {
            "type": "boolean",
            "title": "Patient Reached",
            "default": false
          },
          "responses": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Responses"
          },
          "symptoms_reported": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "type": "array",
            "title": "Symptoms Reported"
          },
          "uncertainties": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Uncertainties"
          },
          "callback_requested": {
            "type": "boolean",
            "title": "Callback Requested",
            "default": false
          },
          "callback_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Callback Time"
          },
          "transcript": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "type": "array",
            "title": "Transcript"
          },
          "summary": {
            "type": "string",
            "title": "Summary",
            "default": ""
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "completed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Completed At"
          }
        },
        "type": "object",
        "required": [
          "call_id",
          "patient_id",
          "campaign_id",
          "outcome",
          "started_at",
          "completed_at"
        ],
        "title": "ConversationResult"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}